task6

+++ 0x00-hello_world/6-size.c
@@ -1,10 +1,9 @@
 #include <stdio.h>
 /**
  * main -print out sizes of data in C
- * * code by bewave ict
- * Return 0 (success)
- */
-
+ * code by bewave ict
+ * Return:always 0
+*/
 int main(void)
 {
 	char a;
@@ -13,10 +12,10 @@ int main(void)
 	long long d;
 	float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu bytes(s)\n", (unsigned int)sizeof(c));
-printf("Size of a long long int: %lu bytes(s)\n", (unsigned int)sizeof(d));
-printf("Size of a float: %lu bytes(s)\n", (unsigned int)sizeof(f));
-return (0);
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
+	return (0);
 }
